+++ server/tests-py/test_openapi.py
@@ -1,3 +1,4 @@
+import collections
 import pytest
 import os
 from validate import check_query_f, check_query, get_conf_f
@@ -52,3 +53,65 @@ class TestOpenAPISpec:
 
     def test_duplicate_field_name(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + '/openapi_get_endpoint_test_duplicate_field_name.yaml', transport)
+
+    def test_inconsistent_schema_openAPI(self, hge_ctx, transport):
+        # export metadata and create a backup
+        st_code, backup_metadata = hge_ctx.v1q(
+            q = {
+                "type": "export_metadata",
+                "args": {}
+            }
+        )
+        assert st_code == 200, backup_metadata
+
+        new_metadata = backup_metadata.copy()
+
+        #create inconsistent metadata
+        inconsistent_query = [collections.OrderedDict([
+            ('name', 'wrong_queries'), 
+            ('definition', collections.OrderedDict([
+                ('queries', [collections.OrderedDict([
+                    ('name', 'random_query'), 
+                    ('query', 'query { random_field_name test_table { random_col_name } }')
+                ])])
+            ]))
+        ])]
+        res_endpoint = [collections.OrderedDict([
+            ("definition", collections.OrderedDict([
+                ("query", collections.OrderedDict([
+                    ("collection_name", "wrong_queries"),
+                    ("query_name", "random_query")
+                ]))
+            ])),
+            ("url", "some_url"),
+            ("methods", ["GET"]),
+            ("name", "wrong_endpoint"),
+            ("comment", None) 
+        ])]
+        new_metadata["query_collections"] = inconsistent_query
+        new_metadata["rest_endpoints"] = res_endpoint
+
+        # apply inconsistent metadata
+        st_code, resp = hge_ctx.v1q(
+            q={
+                "type": "replace_metadata",
+                "version": 2,
+                "args": {
+                    "allow_inconsistent_metadata": True,
+                    "metadata": new_metadata
+                }
+            }
+        )
+        assert st_code == 200, resp
+
+        # check openAPI schema
+        check_query_f(hge_ctx, self.dir() + '/openapi_inconsistent_schema.yaml', transport)
+
+        # revert to old metadata
+        st_code, resp = hge_ctx.v1q(
+            q={
+                "type": "replace_metadata",
+                "args": backup_metadata
+            }
+        )
+        assert st_code == 200, resp